Separate decoding/updating in update_eeprom_db
I think it would be best for the syseeprom daemon to publish the
information to the database. However, for now I want to separate the
decoding and the publishing, so that platform plugins can simply
implement get_eeprom_dict and not worry about publishing. Please let me
know why or why not the publishing is not handled in the daemon instead.

Also created symlink for sonic_eeprom that points to the exact same
files in sonic_platform_base/sonic_eeprom.