[sonic_platform_base] support new-platform-api-based daemons (#48)
[sonic_platform_base/chassis]
1. add _eeprom and get_eeprom into class ChassisBase so that syseepromd can access eeprom related stuff.
2. fix typo (non-asic characters)

[sonic_platform_base/sonic_sfp/sfputilhelper]
For xcvrd, originally we plan to remain the module sfputilbase to provide functions like mapping between logical and physical interfaces. However, SfpUtilBase is an abstract class and can not be instantiated. In that case, a new module sfputilhelper has to be introduced to provide the required interfaces.
1. read_porttab_mappings, which reads port_config.ini and generats mappings
2. get_physical_to_logical
3. get_logical_to_physical
4. is_logical_port

[sonic_platform_base/sfp_base]
support interface get_transceiver_threshold_status