[xcvrd] Fix crashing due to missing fields of some module types (#54)
From SFF-8436 regarding cable length fields:
"A value of zero means that the Module does not support (this mode) or that the length information must be determined from the Module technology."

It is possible that all cable length fields in eeprom are 0, and this would lead to xcvrd crashing due to fetching of this data. Fallback is added in this case.