[sonic_platform_base] reallocate the empty LIST at the constructor of the base class
When we use a mutable object as a default value in the base class, a value that can be modified in place in the subclasses, such as the empty lists of
the chassis base class, and it is most likely unwanted.

And hence, this commit will try to address this issue by allocating new empty
list objects at the constructor of the base class, and the subclass should
rely on the constructor of superclass for this.

As a result, it simplifies the source code, and reduce the maintenance efforts

Signed-off-by: Dante (Kuo-Jung) Su <dante.su@broadcom.com>