[xcvrd] Add enum dependence back; Install 'enum34' conditionally based on Python version (#107)
Add dependence on 'enum' package back to xcvrd (basically reverting most of https://github.com/Azure/sonic-platform-daemons/pull/106). However, in setup.py, we only install the enum34 package if the version of Python we are installing for is < 3.4. Thus, when installing the Python 3 xcvrd package in Python 2.7, the Python 2 version of enum34 will be installed. However, if installing the Python 3 xcvrd package on Python 3.7, enum34 will not be installed, causing xcrvd to import the 'enum' module from the standard library. This should prevent any conflicts which arise when 'enum34' is ever installed on Python versions >= 3.4 by preventing this situation.