[xcvrd] support for integrating Credo Y cable Ports initialization and status updates with xcvrd.  (#105)
* [xcvrd] support for integrating y cable within xcvrd

This PR provides the necessary infrastructure to initialize the Y cable Ports inside SONIC with xcvrd as the platform daemon.
Particularly there are two parts of integration:

While xcvrd initializes , there is within config_db for Y cable presence. This is done by checking the key-value pairs for
presence of mux_cable identifier as a key. Once a Y cable is found to be attached to a port, State DB is updated with
the corresponding data for the Y cable Port.

Once the init process is done, and a Y cable presence is established, A thread is run to periodically monitor changes
to APPL DB MUX_CABLE_COMMAND table for updates, and also one that periodically checks for a change events,  If an update is found, the corresponding changes are done on MUX using
sonic_y_cable package and corresponding changes are updated in STATE_DB

What is the motivation for this PR?
To add the necessary infrastructure for Credo Y cable integration within SONIC

How did you do it?
Added the necessary changes and a new xcvrd_utilities sub directory for utilities of y_cable code.
Reorganized the setup.py and sonix-xcvrd code to this form

sonic-xcvrd/setup.py
sonic-xcvrd/src/init.py
sonic-xcvrd/scripts/xcvrd → sonic-xcvrd/src/xcvrd.py
sonic-xcvrd/src/xcvrd_utilities/init.py
sonic-xcvrd/src/xcvrd_utilities/y_cable_helper.py

Signed-off-by: vaibhav-dahiya <vdahiya@microsoft.com>