[y_cable] Skip ImportError in y_cable.py (#154)
When build python3 xcvrd, it tries to do basic check which will import this y_cable.py. However, not all platform supports python3 API now, so it could cause an issue when importing sonic_platform.platform

We skip the ImportError here to make the builder happy. And this is safe because:
- If any python package is not available, there will be exception when use it
- Vendors know their platform API version, they are responsible to use correct python version when importing this file.