[ci] Test and build packages using Azure Pipelines (#149)
Configure Azure Pipelines to run unit tests on all packages, build all Python wheels and publish the test results, test coverage and resulting wheels.

Note: Azure Pipelines currently has a limitation such that it does not properly support publishing multiple test coverage reports in the same pipeline. If you publish multiple test coverage reports in the same pipeline (which I am doing here), the summary and report is shown for the last task only. Any previously uploaded data is ignored. Hopefully Azure Pipelines will add support for multiple test coverage reports in the near future. See [here](https://docs.microsoft.com/en-us/azure/devops/pipelines/tasks/test/publish-code-coverage-results?view=azure-devops#is-code-coverage-data-merged-when-multiple-files-are-provided-as-input-to-the-task-or-multiple-tasks-are-used-in-the-pipeline) for more info.