[xcvrd] Store mux_cable telemetry data in State DB (#148)
* [xcvrd] adding mux_cable data for sonic-telemetry inside state db

Summary:
This PR provides the necessary infrastructure to initialize the mux_cable info and static tables and post them within state db
as part of xcvrd. The data is posted every 60 secs and streaming telemetry can utilize this info. 
Approach

Added changes in the sonic_xcvrd directory of sonic-platform-daemons
What is the motivation for this PR?

To add the necessary infrastructure for Credo Y cable support for posting streaming telemetry data inside state-db

Signed-off-by: vaibhav-dahiya <vdahiya@microsoft.com>