[xcvrd] Fix crash: If 'dom_capability' not in port_info_dict, insert 'N/A' (#162)
Currently, some vendors are using custom transceiver info parsers which do not yet provide the `dom_capability` field in the results of `get_transceiver_info()`. However, PR https://github.com/Azure/sonic-platform-daemons/pull/72 introduced storing this value to State DB under the assumption that it would always be present. On platforms where this value is not present, it would cause xcvrd to crash (see issue: https://github.com/Azure/sonic-buildimage/issues/6978).

This change will prevent a crash if it is not present, and will in turn save `'N/A'` as the `dom_capability` value in State DB.