[Y-Cable] add the definition inside setup.py to include sonic_y_cable.credo as a package (#211)
Signed-off-by: vaibhav-dahiya <vdahiya@microsoft.com>
Curreently the way python wheel includes packages is all the subdirectories inside a package need to be also included
for them to be a package. Since in the refactor for multiple vendor Y-Cable xcvrd calls these vendor specific modules by including their name which it obtains by reading the eeprom it is required to include them inside setup.py for the wheel to know it has a package.

Signed-off-by: vaibhav-dahiya vdahiya@microsoft.com

Description
add the definition inside setup.py to include sonic_y_cabe.credo as a package

Motivation and Context
This is required by all calling API's for vendor agnostic Y-Cable API's to be called for credo specific cables.

How Has This Been Tested?
Built an image and validated the presence of the pacakge.
Ran the API's on a Arista7050cx3 switch.