[y_cable] add support for manual/standby mode in xcvrd for muxcable; fix download firmware version retrieval logic while download firmware in progress (#220)
This PR supports manual/standby mode support for mux_cable config in addition to auto/active.
currently xcvrd would not create entries for mux_cable if the config is manual, and this PR resolves that problem.
Functionally xcvrd will remain as same if we configure it as active/auto/manual

This PR also addresses the problem when there is download firmware operation in progress and streaming_telemetry/CLI
needs to query the firmware version. In this case we check the status of download_firmware for the port and retrieve the last known value for it from the state DB.
If the download firmware API has failed, we will retrieve the version for cable and just log an error.


Description
Needed for manual/standby config for muxcable operation

Motivation and Context
Required for manual/standby config for muxcable operation
Also addresses the problem when there is download firmware operation in progress and ST thread/CLI
needs to query the firmware version case. In this case we check the status of download_firmware and retrieve the last known value for it for the state DB.

Signed-off-by: vaibhav-dahiya <vdahiya@microsoft.com>