[ycable][credo] Fix the is_link_active API for Credo Ycable (#260)
Signed-off-by: vaibhav-dahiya vdahiya@microsoft.com

Description
Currently is_link_active API was broken because it was not using the target param in the routine to give back the correct link_status of ecah side. With this change it gives the correct result.

Motivation and Context
Required for MUX_CABLE_INFO table in streaming telemetry if the cable is disconnected or the link is down.

redis-cli -n 6 hgetall "MUX_CABLE_INFO|Ethernet16"
"
 9) "link_status_self"
10) "down"
11) "link_status_peer"
12) "up"
13) "link_status_nic"
14) "up"
How Has This Been Tested?
Tested on actual Arista7050cx3 testbed.