[syseeprom] Enable display of vendor extension TLV content
The code that decodes the content of the ONIE syseeprom includes a flag to
enable/disable displaying the content of the vendor extension TLV. This flag is
currently set to 'disable'. Hence the 'show platform syseeprom' command shows
the presence and size of the vendor extension TLV but does not show its
content. This commit sets the flag to 'enable' so that the vendor extension TLV
content is displayed.