Make is_flat_memory more robust (#491)
- Description
Currently, is_flat_memory could return 3 different values:

True when the module is flat memory
False when the module is not flat memory
None when the EEPROM read returns None or empty bytearray
When it returns None, we should not assume that the module is NOT a flat memory. Instead, it would be safer to treat it as a flat memory.

- Motivation and Context
This PR is to make is_flat_memory only returns two state, True or False. And if EEPROM reading returns None or empty bytearray, we should treat the module as a flat memory.

- How Has This Been Tested?
Manual test