[CMIS]The "get_error_description" function should return 'OK' instead of None when there are no errors.
<!-- Provide a general summary of your changes in the Title above -->

#### Description

The "get_error_description" function should return 'OK' instead of 'None' when there are no errors.

#### Motivation and Context
When there are no errors on the cable, get_error_description currently returns 'None'. In this case, we should return 'OK' so that the CLI (show interface transceiver error-description) provides a proper output.

#### How Has This Been Tested?
Call this API directly on different cables with different error statuses.

#### Additional Information (Optional)