Restore optoe page to default 0 for active cables (#549)
<!-- Provide a general summary of your changes in the Title above -->

#### Description
Restore the optoe's page to page ZERO before reading the Lower page > 127

#### Motivation and Context
When the module is busy with CDB command, the optoe kernel driver at the end of user page read request, tries to restore the page to Page ZERO and if this write fails because the module could be busy with CDB command (since these modules donot support background CDB mode). Because of this when Xcvrd tries to read offset (> 127) in Lower page, since optoe does not restore the page to 0 for any Lower page access, the page read to these offset will result in wrong page to be read (because of previous page select write failure) and Xcvrd parsing fails causing crash.

Now in optoe EEPROM read API we ensure to restore the Page to ZERO page if the Page select byte is non-zero

#### How Has This Been Tested?
TBD:
1. Test on Arista switch
2. Test on Cisco switch

#### Additional Information (Optional)