Fix health check for SSD vendors: add a parser for ATP, and add a generic health ID for other brands (#598)
#### Description

Fix health check for SSD vendors: add a parser for ATP, and add a generic health ID for other brands.
Each vendor stores health information in different SMART attributes.
ATP stores it in attribute ID 248, we add a parser for it.
We also have SSDs use Attribute ID 231 and it is commonly used, so add it in the generic parser.
Skip obtaining vendor SSD info for ATP and Virtium NVMe SSD because they are handle by parse_generic_ssd_info and parse_vendor_ssd_info will overwrite data with N/A.
Add unit test cases for ATP SATA/NVMe SSD.

#### Motivation and Context

`show platform ssdhealth` shows N/A health for some qualified SSDs.

#### Back port request
- [x] 202412
- [x] 202505

#### How Has This Been Tested?

We have tested the code change on DUTs with different SSDs including all the qualified SSDs that show N/A in health and also on the ones that worked fine before.