Fix log prefix hijack in c_cmis.py
If any code imports c_cmis.py (indirectly for the most part) and has a
logger session open already, their existing session will be re-opened
with a new log tag, CCmisApi which is not desired.

To fix this we match the logging in c_cmis.py with logging in
neighboring files, it doesn't do any syslogging anymore.

Libraries (rather than executables) probably shouldn't be syslogging on
their own, and they certainly shouldn't hijack the log tag of the
including executable.